feat: rename study app to research
