+++ backend/research/queries.py
@@ -3,8 +3,8 @@ from graphene import Field, List, ObjectType, ResolveInfo, ID, String, NonNull
 
 from django.db.models import QuerySet
 
-from study.models import Study
-from study.types.StudyType import StudyType
+from research.models import Study
+from research.types.StudyType import StudyType
 from main.models import MRPermission
 
 
@@ -39,5 +39,5 @@ class StudyQuery(ObjectType):
     @staticmethod
     def resolve_studies(root, info: ResolveInfo, mr_permission: str) -> QuerySet[Study]:
         return StudyType.get_queryset(
-            Study.objects, info, mr_permission=mr_permission
+            Study.objects, info
         ).accessible_objects(info.context.user, mr_permission)
